Update namespace for cops
It's now Rubocop::Cop::RootCops::Class. This is more in line with how a
real gem should be formed
