+++ lib/rubocop/cop/root_cops/raise_i18n.rb
@@ -1,0 +1,23 @@
+module RuboCop
+  module Cop
+    module RootCops
+      class RaiseI18n < Cop
+        MESSAGE = "Use raise with I18n instead of hard-coded strings".freeze
+
+        def on_send(node)
+          return unless node.command?(:raise)
+
+          node.child_nodes.each do |child|
+            add_offense(child, :location => :expression, :message => MESSAGE) if _str?(child)
+          end
+        end
+
+        private
+
+        def _str?(node)
+          node.str_type? || node.dstr_type? || node.xstr_type?
+        end
+      end
+    end
+  end
+end
